requirements.txt
flask==3.0.3
pandas==2.2.2
numpy==1.26.4
mistralai==1.2.3
pydantic==2.9.0
python-dotenv==1.0.1
rdkit-pypi==2022.9.5
matplotlib==3.9.0
